sfbpf: Remove sfbpf entirely and refresh sfbpf_dlt.h as daq_dlt.h
As much fun as it was to maintain a fork of libpcap's BPF
implementation, I can't really justify it anymore.  If modules want to
implement BPF filtering, they can use libpcap to do so (see the previous
change to AFPacket).  Netmap and IPFW will maintain references to SFBPF
until they are rewritten for the NG API.