api: Add total instances and instance ID to instance configurations
Defaults for both are 0 (unspecified). DAQTest has been updated to pass
both when the number of threads has been explicity requested via the -z
option.  There is no API requirement that either of them be set, but if
both are set, the instance ID must be less than or equal to the total.